[APM] [User Experience] remove uiFilters from setupRequest (#109908)
* remove uiFilters from setupRequest

* adjust get_buckets test

* update getAnomalySeries
