+++ x-pack/plugins/apm/server/lib/errors/distribution/get_buckets.test.ts
@@ -44,7 +44,6 @@ describe('get buckets', () => {
             get: () => 'myIndex',
           }
         ) as APMConfig,
-        uiFilters: {},
         indices: {
           /* eslint-disable @typescript-eslint/naming-convention */
           'apm_oss.sourcemapIndices': 'apm-*',
