+++ x-pack/plugins/apm/server/lib/rum_client/get_page_load_distribution.ts
@@ -8,7 +8,8 @@
 import { TRANSACTION_DURATION } from '../../../common/elasticsearch_fieldnames';
 import { getRumPageLoadTransactionsProjection } from '../../projections/rum_page_load_transactions';
 import { mergeProjection } from '../../projections/util/merge_projection';
-import { Setup, SetupTimeRange } from '../helpers/setup_request';
+import { SetupTimeRange } from '../helpers/setup_request';
+import { SetupUX } from '../../routes/rum_client';
 
 export const MICRO_TO_SEC = 1000000;
 
@@ -64,7 +65,7 @@ export async function getPageLoadDistribution({
   maxPercentile,
   urlQuery,
 }: {
-  setup: Setup & SetupTimeRange;
+  setup: SetupUX & SetupTimeRange;
   minPercentile?: string;
   maxPercentile?: string;
   urlQuery?: string;
@@ -176,7 +177,7 @@ const getPercentilesDistribution = async ({
   minDuration,
   maxDuration,
 }: {
-  setup: Setup & SetupTimeRange;
+  setup: SetupUX & SetupTimeRange;
   minDuration: number;
   maxDuration: number;
 }) => {
