+++ x-pack/plugins/apm/server/lib/rum_client/get_pl_dist_breakdown.ts
@@ -8,7 +8,8 @@
 import { getRumPageLoadTransactionsProjection } from '../../projections/rum_page_load_transactions';
 import { ProcessorEvent } from '../../../common/processor_event';
 import { mergeProjection } from '../../projections/util/merge_projection';
-import { Setup, SetupTimeRange } from '../helpers/setup_request';
+import { SetupTimeRange } from '../helpers/setup_request';
+import { SetupUX } from '../../routes/rum_client';
 import {
   CLIENT_GEO_COUNTRY_ISO_CODE,
   USER_AGENT_DEVICE,
@@ -44,7 +45,7 @@ export const getPageLoadDistBreakdown = async ({
   breakdown,
   urlQuery,
 }: {
-  setup: Setup & SetupTimeRange;
+  setup: SetupUX & SetupTimeRange;
   minPercentile: number;
   maxPercentile: number;
   breakdown: string;
