+++ x-pack/plugins/apm/server/lib/rum_client/get_url_search.ts
@@ -6,7 +6,8 @@
  */
 
 import { mergeProjection } from '../../projections/util/merge_projection';
-import { Setup, SetupTimeRange } from '../helpers/setup_request';
+import { SetupTimeRange } from '../helpers/setup_request';
+import { SetupUX } from '../../routes/rum_client';
 import { getRumPageLoadTransactionsProjection } from '../../projections/rum_page_load_transactions';
 import {
   TRANSACTION_DURATION,
@@ -18,7 +19,7 @@ export async function getUrlSearch({
   urlQuery,
   percentile,
 }: {
-  setup: Setup & SetupTimeRange;
+  setup: SetupUX & SetupTimeRange;
   urlQuery?: string;
   percentile: number;
 }) {
