+++ x-pack/plugins/apm/server/lib/rum_client/get_web_core_vitals.ts
@@ -7,7 +7,8 @@
 
 import { getRumPageLoadTransactionsProjection } from '../../projections/rum_page_load_transactions';
 import { mergeProjection } from '../../projections/util/merge_projection';
-import { Setup, SetupTimeRange } from '../helpers/setup_request';
+import { SetupTimeRange } from '../helpers/setup_request';
+import { SetupUX } from '../../routes/rum_client';
 import {
   CLS_FIELD,
   FCP_FIELD,
@@ -21,7 +22,7 @@ export async function getWebCoreVitals({
   urlQuery,
   percentile = 50,
 }: {
-  setup: Setup & SetupTimeRange;
+  setup: SetupUX & SetupTimeRange;
   urlQuery?: string;
   percentile?: number;
 }) {
