+++ x-pack/plugins/apm/server/lib/rum_client/has_rum_data.ts
@@ -5,7 +5,8 @@
  * 2.0.
  */
 
-import { Setup, SetupTimeRange } from '../helpers/setup_request';
+import { SetupTimeRange } from '../helpers/setup_request';
+import { SetupUX } from '../../routes/rum_client';
 import {
   SERVICE_NAME,
   TRANSACTION_TYPE,
@@ -17,7 +18,7 @@ import { TRANSACTION_PAGE_LOAD } from '../../../common/transaction_types';
 export async function hasRumData({
   setup,
 }: {
-  setup: Setup & Partial<SetupTimeRange>;
+  setup: SetupUX & Partial<SetupTimeRange>;
 }) {
   try {
     const { start, end } = setup;
