+++ x-pack/plugins/apm/server/lib/transactions/get_anomaly_data/index.ts
@@ -22,6 +22,7 @@ export async function getAnomalySeries({
   serviceName,
   transactionType,
   transactionName,
+  kuery,
   setup,
   logger,
 }: {
@@ -29,6 +30,7 @@ export async function getAnomalySeries({
   serviceName: string;
   transactionType: string;
   transactionName?: string;
+  kuery: string;
   setup: Setup & SetupTimeRange;
   logger: Logger;
 }) {
@@ -50,13 +52,8 @@ export async function getAnomalySeries({
     return undefined;
   }
 
-  // Don't fetch anomalies if uiFilters are applied. This filters out anything
-  // with empty values so `kuery: ''` returns false but `kuery: 'x:y'` returns true.
-  const hasUiFiltersApplied =
-    Object.entries(setup.uiFilters).filter(([_key, value]) => !!value).length >
-    0;
-
-  if (hasUiFiltersApplied) {
+  // Don't fetch anomalies if kuery is present
+  if (kuery) {
     return undefined;
   }
